+++ src/test/converter2/issues/gh3017.ts
@@ -1,0 +1,8 @@
+export class GH3017 {
+    /** @hidden */
+    constructor() {}
+    #hashPriv = 1;
+    private priv = 2;
+    protected prot = 3;
+    public pub = 4;
+}